Fiz a função de iluminação
Agora o jogo tem um sol que fica iluminando, mas falta arrumar a normal de tudo o que está sendo desenhado

+++ include/map.h
@@ -22,7 +22,7 @@ extern square mapCells[MAP_SIZE][MAP_SIZE]; // A --- C
                                             // |     |
                                             // B --- D
 
-extern int heightMatrix[MAP_SIZE][MAP_SIZE];
+extern float heightMatrix[MAP_SIZE][MAP_SIZE];
 
 void drawMap();
